Some refactoring

+++ app/routes/tournament._index.tsx
@@ -1,11 +1,9 @@
 import type { ActionArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
-import { Link, useLoaderData } from "@remix-run/react";
-import { useCallback, useState } from "react";
-import { Tournaments } from "~/components/templates/tournaments";
+import { useLoaderData } from "@remix-run/react";
+import { TournamentsTemplate } from "~/components/templates/TournamentsTemplate";
 import { TournamentService } from "~/services";
 import type { Dto } from "~/types/base";
-import type { Tournament } from "~/types/tournament";
 
 export const loader = async () => {
   var tournaments = await TournamentService.getAllTournaments();
@@ -16,13 +14,14 @@ export const loader = async () => {
 
 export async function action({ request }: ActionArgs) {
   const bodyOb = (await request.json()) as Dto<{}>;
-
-  TournamentService.deleteTournament(bodyOb._id);
+  if (request.method === "DELETE") {
+    TournamentService.deleteTournament(bodyOb._id);
+  }
   return null;
 }
 
 export default function TournamentIndex() {
   const { tournaments } = useLoaderData<typeof loader>();
 
-  return <Tournaments data={tournaments} />;
+  return <TournamentsTemplate data={tournaments} />;
 }
